Move exceptions into exception module.
