+++ acceptancetests/jujupy/client.py
@@ -24,10 +24,7 @@ from contextlib import (
     contextmanager,
     )
 from copy import deepcopy
-from datetime import (
-    datetime,
-    timedelta,
-    )
+from datetime import datetime
 import errno
 from itertools import chain
 import json
@@ -49,6 +46,35 @@ from jujupy.configuration import (
     get_bootstrap_config_path,
     get_juju_home,
     get_selected_environment,
+    )
+from jujupy.exceptions import (
+    AgentError,
+    AgentUnresolvedError,
+    AgentsNotStarted,
+    AppError,
+    ApplicationsNotStarted,
+    AuthNotAccepted,
+    CannotConnectEnv,
+    ErroredUnit,
+    HookFailedError,
+    InstallError,
+    InvalidEndpoint,
+    NameNotAccepted,
+    NoActiveControllers,
+    NoActiveModel,
+    NoProvider,
+    MachineError,
+    ProvisioningError,
+    SoftDeadlineExceeded,
+    StatusError,
+    StatusNotMet,
+    StatusTimeout,
+    StuckAllocatingError,
+    TypeNotAccepted,
+    UnitError,
+    VersionsNotUpdated,
+    VotingNotEnabled,
+    WorkloadsNotReady,
     )
 from jujupy.utility import (
     get_timeout_path,
@@ -92,46 +118,6 @@ _DEFAULT_BUNDLE_TIMEOUT = 3600
 log = logging.getLogger("jujupy")
 
 
-class StatusTimeout(Exception):
-    """Raised when 'juju status' timed out."""
-
-
-class SoftDeadlineExceeded(Exception):
-    """Raised when an overall client operation takes too long."""
-
-    def __init__(self):
-        super(SoftDeadlineExceeded, self).__init__(
-            'Operation exceeded deadline.')
-
-
-class NoProvider(Exception):
-    """Raised when an environment defines no provider."""
-
-
-class TypeNotAccepted(Exception):
-    """Raised when the provided type was not accepted."""
-
-
-class NameNotAccepted(Exception):
-    """Raised when the provided name was not accepted."""
-
-
-class InvalidEndpoint(Exception):
-    """Raised when the provided endpoint was deemed invalid."""
-
-
-class AuthNotAccepted(Exception):
-    """Raised when the provided auth was not accepted."""
-
-
-class NoActiveModel(Exception):
-    """Raised when no active model could be found."""
-
-
-class NoActiveControllers(Exception):
-    """Raised when no active environment could be found."""
-
-
 def get_timeout_prefix(duration, timeout_path=None):
     """Return extra arguments to run a command with a timeout."""
     if timeout_path is None:
@@ -160,22 +146,6 @@ def parse_new_state_server_from_error(error):
     return None
 
 
-class ErroredUnit(Exception):
-
-    def __init__(self, unit_name, state):
-        msg = '%s is in state %s' % (unit_name, state)
-        Exception.__init__(self, msg)
-        self.unit_name = unit_name
-        self.state = state
-
-
-class UpgradeMongoNotSupported(Exception):
-
-    def __init__(self):
-        super(UpgradeMongoNotSupported, self).__init__(
-            'This client does not support upgrade-mongo')
-
-
 Machine = namedtuple('Machine', ['machine_id', 'info'])
 
 
@@ -189,49 +159,6 @@ def coalesce_agent_status(agent_item):
     if state is None:
         state = 'no-agent'
     return state
-
-
-class CannotConnectEnv(subprocess.CalledProcessError):
-
-    def __init__(self, e):
-        super(CannotConnectEnv, self).__init__(e.returncode, e.cmd, e.output)
-
-
-class StatusNotMet(Exception):
-
-    _fmt = 'Expected status not reached in {env}.'
-
-    def __init__(self, environment_name, status):
-        self.env = environment_name
-        self.status = status
-
-    def __str__(self):
-        return self._fmt.format(env=self.env)
-
-
-class AgentsNotStarted(StatusNotMet):
-
-    _fmt = 'Timed out waiting for agents to start in {env}.'
-
-
-class VersionsNotUpdated(StatusNotMet):
-
-    _fmt = 'Some versions did not update.'
-
-
-class WorkloadsNotReady(StatusNotMet):
-
-    _fmt = 'Workloads not ready in {env}.'
-
-
-class ApplicationsNotStarted(StatusNotMet):
-
-    _fmt = 'Timed out waiting for applications to start in {env}.'
-
-
-class VotingNotEnabled(StatusNotMet):
-
-    _fmt = 'Timed out waiting for voting to be enabled in {env}.'
 
 
 class JujuData:
@@ -606,74 +533,6 @@ class JujuData:
 
     def __ne__(self, other):
         return not self == other
-
-
-class StatusError(Exception):
-    """Generic error for Status."""
-
-    recoverable = True
-
-    # This has to be filled in after the classes are declared.
-    ordering = []
-
-    @classmethod
-    def priority(cls):
-        """Get the priority of the StatusError as an number.
-
-        Lower number means higher priority. This can be used as a key
-        function in sorting."""
-        return cls.ordering.index(cls)
-
-
-class MachineError(StatusError):
-    """Error in machine-status."""
-
-    recoverable = False
-
-
-class ProvisioningError(MachineError):
-    """Machine experianced a 'provisioning error'."""
-
-
-class StuckAllocatingError(MachineError):
-    """Machine did not transition out of 'allocating' state."""
-
-    recoverable = True
-
-
-class UnitError(StatusError):
-    """Error in a unit's status."""
-
-
-class HookFailedError(UnitError):
-    """A unit hook has failed."""
-
-    def __init__(self, item_name, msg):
-        match = re.search('^hook failed: "([^"]+)"$', msg)
-        if match:
-            msg = match.group(1)
-        super(HookFailedError, self).__init__(item_name, msg)
-
-
-class InstallError(HookFailedError):
-    """The unit's install hook has failed."""
-
-    recoverable = True
-
-
-class AppError(StatusError):
-    """Error in an application's status."""
-
-
-class AgentError(StatusError):
-    """Error in a juju agent."""
-
-
-class AgentUnresolvedError(AgentError):
-    """Agent error has not recovered in a reasonable time."""
-
-    # This is the time limit set by IS for recovery from an agent error.
-    a_reasonable_time = timedelta(minutes=5)
 
 
 StatusError.ordering = [
